+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Add `equalsCString`.
 * Add `stripCStringPrefix`.
 * Add `equalsLatin8`.
+* Add `emptyPinned`.
+* Add `concatPinned` to `Data.Bytes.Chunks`.
 
 ## 0.2.1.0 -- 2020-01-22